Vortex Elements

The library currently has four built-in vortex types:

Vortex.Point

Vortex.Blob

Vortex.Sheet

Vortex.Plate (at the moment, there can only be one plate in the fluid at at time)

Most functions in the library that act on vortex elements can take either a single vortex element, or a collection of elements. These collections can be represented as an array or a tuple. Arrays should be used when the elements are the same type, for example:

julia> points = Vortex.Point.(rand(Complex128, 5), rand(5))
5-element Array{VortexModel.Vortex.Points.Point,1}:
 Point Vortex: z = 0.236 + 0.347im, Γ = 0.556
 Point Vortex: z = 0.313 + 0.008im, Γ = 0.437
 Point Vortex: z = 0.489 + 0.211im, Γ = 0.425
 Point Vortex: z = 0.952 + 1.0im, Γ = 0.773
 Point Vortex: z = 0.252 + 0.987im, Γ = 0.281

julia> Vortex.impulse(points)
1.3362266530178137 - 1.2821936908564113im

julia> blobs = [Vortex.Blob(rand(Complex128), rand(), 0.1) for i in 1:5]
5-element Array{VortexModel.Vortex.Blobs.Blob,1}:
 Vortex Blob: z = 0.209 + 0.251im, Γ = 0.02, δ = 0.1
 Vortex Blob: z = 0.288 + 0.86im, Γ = 0.077, δ = 0.1
 Vortex Blob: z = 0.64 + 0.874im, Γ = 0.279, δ = 0.1
 Vortex Blob: z = 0.751 + 0.645im, Γ = 0.078, δ = 0.1
 Vortex Blob: z = 0.848 + 0.086im, Γ = 0.553, δ = 0.1

julia> Vortex.impulse(blobs)
0.41217890550975256 - 0.7325028967929701im

Knowing that every element has the same type allows the compiler to perform more aggressive optimizations. Tuples are used when we want to mix and match different vortex types. For example:

julia> sys = (points, blobs);

julia> Vortex.impulse(sys)
1.7484055585275664 - 2.0146965876493814im

This rest of this page documents the data types that represent these elements and some key functions that act on them. For more detailed examples, please refer to the Jupyter notebooks.

Built-in Vortex Types

VortexModel.Vortex.Points.Point — Type.

Vortex.Point <: Vortex.PointSource

An immutable structure representing a point vortex

Fields

z: position

Γ: circulation

source

VortexModel.Vortex.Blobs.Blob — Type.

Vortex.Blob <: Vortex.PointSource

An immutable structure representing a vortex blob

Fields

z: position

Γ: circulation

δ: blob radius

source

VortexModel.Vortex.Sheets.Sheet — Type.

Vortex.Sheet <: Vortex.CompositeSource

A vortex sheet represented by vortex blob control points

Fields

blobs: the underlying array of vortex blobs

Γs: the cumulated sum of circulation starting from the first control point

δ: the blob radius of all the vortex blobs

zs: a mapped array that accesses the position of each control point

Constructors:

Sheet(blobs, Γs, δ)

Sheet(zs, Γs, δ) where zs is an array of positions for the control points

source

VortexModel.Vortex.Plates.Plate — Type.

Vortex.Plate <: VortexCompositeSource

An infinitely thin, flat plate, represented as a bound vortex sheet

Fields

L

chord length

c

centroid

α

centroid velocity

Γ

total circulation

N

number of control points

ss

normalized positions (within [-1, 1]) of the control points

zs

control point coordinates

A

Chebyshev coefficients of the normal component of velocity induced along the plate by ambient vorticity

C

Chebyshev coefficients of the velocity induced along the plate by ambient vorticity

B₀

zeroth Chebyshev coefficient associated with body motion

B₁

first Chebyshev coefficient associated with body motion

dct!

Preplanned DCT used to perform the discrete Chebyshev transform

Constructors

Plate(N, L, c, α)

source

Vortex Properties

VortexModel.Vortex.position — Function.

Vortex.position(src::PointSource)

Returns the complex position of a PointSource type vortex element This is a required method for all subtypes of PointSource.

Example

julia> point = Vortex.Point(1.0 + 0.0im, 1.0);

julia> Vortex.position(point)
1.0 + 0.0im

julia> points = Vortex.Point.([1.0im, 2.0im], 1.0);

julia> Vortex.position.(points)
2-element Array{Complex{Float64},1}:
 0.0+1.0im
 0.0+2.0im

source

VortexModel.Vortex.circulation — Function.

Vortex.circulation(src)

Returns the total circulation contained in src This is a required method for all vortex types.

Example

julia> points = Vortex.Point.([1.0im, 2.0im], [1.0, 2.0]);

julia> Vortex.circulation(points[1])
1.0

julia> Vortex.circulation(points)
3.0

julia> Vortex.circulation.(points)
2-element Array{Float64,1}:
 1.0
 2.0

source

VortexModel.Vortex.impulse — Function.

Vortex.impulse(src)

Return the aerodynamic impulse of src about (0,0):

\[P := \int \boldsymbol{x} \times \boldsymbol{\omega}\,\mathrm{d}A.\]

This is a required method for all vortex types.

Example

julia> sys = (Vortex.Point(1.0im, π), Vortex.Blob(1.0im, -π, 0.1));

julia> Vortex.impulse(sys[1])
3.141592653589793 + 0.0im

julia> Vortex.impulse(sys)
0.0 + 0.0im

source

VortexModel.Vortex.advect — Function.

advect(src::PointSource, velocity::Complex128, Δt)

Return a new vortex element that represents src advected by velocity over Δt If this method is implemented by any type T <: PointSource, then an array of type AbstractArray{T} can be passed in the first two arguments of advect!.

Example

julia> point = Vortex.Point(1.0 + 0.0, 1.0);

julia> advect(point, 1.0im, 1e-2)
Point Vortex: z = 1.0 + 0.01im, Γ = 1.0

source

VortexModel.Vortex.advect! — Function.

advect!(srcs₊, srcs₋, vels, Δt)

Moves the elements in srcs₋ by their corresponding velocity in vels over the interval Δt and store the results in src₊ srcs₋ and srcs₊ can be either a array of vortex elements or a tuple.

Example

julia> points₋ = [Vortex.Point(x + 0im, 1.0) for x in 1:5];

julia> points₊ = Vector{Vortex.Point}(5);

julia> vels = [ y*im for y in 1.0:5 ];

julia> advect!(points₊, points₋, vels, 1e-2)

julia> points₊
5-element Array{VortexModel.Vortex.Points.Point,1}:
 Point Vortex: z = 1.0 + 0.01im, Γ = 1.0
 Point Vortex: z = 2.0 + 0.02im, Γ = 1.0
 Point Vortex: z = 3.0 + 0.03im, Γ = 1.0
 Point Vortex: z = 4.0 + 0.04im, Γ = 1.0
 Point Vortex: z = 5.0 + 0.05im, Γ = 1.0

source

Methods on Vortex Sheets

VortexModel.Vortex.Sheets.append_segment! — Function.

Vortex.Sheets.append_segment!(sheet::Sheet, z, Γ)

Append a new segment with circulation Γ extending from the end of the sheet to z.

Example

julia> sheet = Vortex.Sheet(0:0.1:1, 0.0:10, 0.2)
Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2

julia> sheet.blobs[end]
Vortex Blob: z = 1.0 + 0.0im, Γ = 0.5, δ = 0.2

julia> Vortex.Sheets.append_segment!(sheet, 1.1, 2.0)

julia> sheet
Vortex Sheet: L ≈ 1.1, Γ = 12.0, δ = 0.2

julia> sheet.blobs[end]
Vortex Blob: z = 1.1 + 0.0im, Γ = 1.0, δ = 0.2

source

VortexModel.Vortex.Sheets.truncate! — Function.

Vortex.Sheets.truncate!(sheet, n::Int)

Remove segments 0:n from sheet, and return the circulation in those segments.

Example

julia> sheet = Vortex.Sheet(0:0.1:1, 0.0:10, 0.2)
Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2

julia> Vortex.Sheets.truncate!(sheet, 5)
4.0

source

VortexModel.Vortex.Sheets.redistribute_points! — Function.

Vortex.Sheets.redistribute_points!(sheet, zs, Γs)

Returns the modified sheet with replacement control points at positions zs and strength Γs.

julia> sheet = Vortex.Sheet(0:0.1:1, 0.0:10, 0.2)
Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2

julia> sys = (sheet,)
(Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2,)

julia> Vortex.Sheets.redistribute_points!(sheet, 0:0.2:2, 0.0:0.5:5)
Vortex Sheet: L ≈ 2.0, Γ = 5.0, δ = 0.2

julia> sys[1]
Vortex Sheet: L ≈ 2.0, Γ = 5.0, δ = 0.2

source

VortexModel.Vortex.Sheets.remesh — Function.

Vortex.Sheets.remesh(sheet, Δs::Float64 , params::Tuple = ())

Uniformly redistribute the control points of the sheet to have a nominal spacing of Δs. Material quantities that should be redistributed along with the control points can be passed in as elements of params.

Returns the tuple (z₌, Γ₌, L [, p₌]) where

z₌ is an array with the positions of the uniformly distributed points

Γ₌ is circulation interpolated onto z₌

L is total length of the sheet

p₌ is a tuple containing the material quantities from params interpolated onto z₌

Example

julia> sheet = Vortex.Sheet(0:0.1:1, 0.0:10, 0.2)
Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2

julia> age = collect(10.0:-1:0);

julia> Vortex.Sheets.remesh(sheet, 0.2, (age, ))
(Complex{Float64}[0.0+0.0im, 0.25+0.0im, 0.5+0.0im, 0.75+0.0im, 1.0+0.0im], [0.0, 2.5, 5.0, 7.5, 10.0], 1.0, ([10.0, 7.5, 5.0, 2.5, 0.0],))

source

VortexModel.Vortex.Sheets.remesh! — Function.

Vortex.Sheets.remesh!(sheet::Sheet, Δs::Float64, params::Tuple = ())

Same as Vortex.Sheets.remesh, except sheet is replaced internally by a uniformly interpolated control points. Returns the tuple (sheet, L, p₌) where

sheet is the modified sheet

L is total length of the sheet

p₌ is a tuple containing the material quantities from params interpolated onto the new control points of sheet

julia> sheet = Vortex.Sheet(0:0.1:1, 0.0:10, 0.2)
Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2

julia> age = collect(10.0:-1:0);

julia> Vortex.Sheets.remesh!(sheet, 0.2, (age,));

julia> Vortex.position.(sheet.blobs)
5-element Array{Complex{Float64},1}:
  0.0+0.0im
 0.25+0.0im
  0.5+0.0im
 0.75+0.0im
  1.0+0.0im

julia> age
5-element Array{Float64,1}:
 10.0
  7.5
  5.0
  2.5
  0.0

source

VortexModel.Vortex.Sheets.split! — Function.

Vortex.Sheets.split!(sheet, n::Int)

Remove segments 0:n from sheet, and return those segments as a new sheet.

Example

julia> sheet = Vortex.Sheet(0:0.1:1, 0.0:10, 0.2)
Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2

julia> sheet₋ = Vortex.Sheets.split!(sheet, 5)
Vortex Sheet: L ≈ 0.4, Γ = 4.0, δ = 0.2

julia> sheet
Vortex Sheet: L ≈ 0.6, Γ = 6.0, δ = 0.2

source

VortexModel.Vortex.Sheets.filter! — Function.

Vortex.Sheets.filter!(sheet, Δs, Δf[, params])

Redistribute and filter the control points of a vortex sheet

Arguments

sheet: the vortex sheet to be modified

Δs: the nominal spacing between the uniform points

Δf: the minimum length scale that the filter should allow to pass through

params: an optional tuple of vectors containing material properties

Returns

If params is passed in, then its vectors will be overwritten by their interpolated values on the new control points, and the function returns the tuple (sheet, params). Otherwise, it returns (sheet, ())

source

VortexModel.Vortex.Sheets.filter_position! — Function.

filter_position!(s, Δf, L = arclength(z₌))

Filter out any length scales in s that is smaller than Δf, storing the result back in s. s can be either a vector of complex positions, or a Vortex.Sheet.

source

VortexModel.Vortex.Sheets.arclength — Function.

arclength(s)

Compute the polygonal arc length of s, where s can be either an vector of complex numbers or a Vortex.Sheet.

Example

```jldoctest julia> sheet = Vortex.Sheet(0:0.1:1, 0.0:10, 0.2) Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2

julia> Vortex.Sheets.arclength(sheet) 1.0

source

VortexModel.Vortex.Sheets.arclengths — Function.

arclengths(s)

Cumulative sum of the polygonal arc length of s, where s can be either an vector of complex numbers or a Vortex.Sheet.

Example

julia> sheet = Vortex.Sheet(0:0.1:1, 0.0:10, 0.2)
Vortex Sheet: L ≈ 1.0, Γ = 10.0, δ = 0.2

julia> Vortex.Sheets.arclengths(sheet)
11-element Array{Float64,1}:
 0.0
 0.1
 0.2
 0.3
 0.4
 0.5
 0.6
 0.7
 0.8
 0.9
 1.0

source

Methods on Plates

VortexModel.Vortex.Plates.enforce_no_flow_through! — Function.

enforce_no_flow_through!(p::Plate, motion, elements)

Update the plate, p, to enforce the no-flow-through condition given ambient vortex elements, elements, and while moving with kinematics specified by motion.

Example

julia> plate = Vortex.Plate(128, 2.0, 0.0, π/3)
Plate: N = 128, L = 2.0, c = 0.0 + 0.0im, α = 60.0ᵒ
       LESP = 0.0, TESP = 0.0

julia> motion = allocate_velocity(plate); motion.ċ = 1.0;

julia> point = Vortex.Point(0.0 + 2im, 1.0);

julia> Vortex.enforce_no_flow_through!(plate, motion, point)

julia> plate
Plate: N = 128, L = 2.0, c = 0.0 + 0.0im, α = 60.0ᵒ
       LESP = 1.27, TESP = -1.93

source

VortexModel.Vortex.Plates.vorticity_flux — Function.

vorticity_flux(p::Plate, v₁, v₂,
               lesp = 0.0, tesp = 0.0,
               ∂C₁ = Vector{Complex128}(plate.N),
               ∂C₂ = Vector{Complex128}(plate.N))

Return strengths of new vortex elements that satisfies edge suction parameters. For a given edge, if the current suction parameter is less than the criticial suction parameter, then no vorticity is released. If it is higher, however, vorticity will be released so that the suction parameter equals the critical value.

Arguments

p: the plate

v₁, v₂: the vortex elements (with unit circulation) that the vorticity flux is going into

lesp, tesp: the critical leading and trailing edge suction parameters we want to enforce. By default, both parameters are set to 0.0 to enforce the Kutta condition on both edges. We can disable vortex shedding from an edge by setting the its critical suction parameter to Inf

Returns

Γ₁, Γ₂: the strengths that the vortex element should have in order to satisfy the edge suction parameters

∂C₁, ∂C₂: Chebyshev coefficients of the normal velocity induced by the vortex elements Instead of running enforce_bc! with the new vortex elements, we can use this matrix to directly update the Chebyshev coefficients associated with the bound vortex sheet without recomputing all the velocities.

Example

Enforcing the trailing edge Kutta condition with an point vortex at negative infinity:

julia> plate = Vortex.Plate(128, 2.0, 0.0, π/6)
Plate: N = 128, L = 2.0, c = 0.0 + 0.0im, α = 30.0ᵒ
       LESP = 0.0, TESP = 0.0

julia> motion = allocate_velocity(plate);

julia> motion.ċ = 1.0;

julia> Vortex.enforce_no_flow_through!(plate, motion, ())

julia> point = Vortex.Point(-Inf, 1.0);

julia> _, Γ, _, _ = Vortex.vorticity_flux(plate, (), point,  Inf);

julia> Γ # should equal -πULsin(α) = -π
-3.1415926535897927

source

VortexModel.Vortex.Plates.vorticity_flux! — Function.

vorticity_flux!(p::Plate, v₁, v₂,
                lesp = 0.0, tesp = 0.0,
                ∂C₁ = Vector{Complex128}(plate.N),
                ∂C₂ = Vector{Complex128}(plate.N))

In-place version of vorticity_flux, except instead of just returning the possible changes in plate Chebyshev coefficients, we modify plate.C with those changes so that no-flow-through is enforced in the presence of v₁ and v₂ with strengths that satisfy the suction parameters.

source

VortexModel.Vortex.Plates.bound_circulation — Function.

bound_circulation(plate[, s])

Compute the bound circulation on the plate.

s can be either a single normalized arc length coordinate (between -1 and 1), or a whole array of coordinates.

source

VortexModel.Vortex.Plates.bound_circulation! — Function.

bound_circulation!(γs, plate[, ss])

Compute the bound circulation of plate and store it in γs.

If an array, ss, with normalized arc length coordinates is omitted, then the circulation will be computed at the plate's Chebyshev nodes.

source

Index

VortexModel.Vortex.Blobs.Blob

VortexModel.Vortex.Plates.Plate

VortexModel.Vortex.Points.Point

VortexModel.Vortex.Sheets.Sheet

VortexModel.Vortex.Plates.bound_circulation

VortexModel.Vortex.Plates.bound_circulation!

VortexModel.Vortex.Plates.enforce_no_flow_through!

VortexModel.Vortex.Plates.vorticity_flux

VortexModel.Vortex.Plates.vorticity_flux!

VortexModel.Vortex.Sheets.append_segment!

VortexModel.Vortex.Sheets.arclength

VortexModel.Vortex.Sheets.arclengths

VortexModel.Vortex.Sheets.filter!

VortexModel.Vortex.Sheets.filter_position!

VortexModel.Vortex.Sheets.redistribute_points!

VortexModel.Vortex.Sheets.remesh

VortexModel.Vortex.Sheets.remesh!

VortexModel.Vortex.Sheets.split!

VortexModel.Vortex.Sheets.truncate!

VortexModel.Vortex.advect

VortexModel.Vortex.advect!

VortexModel.Vortex.circulation

VortexModel.Vortex.impulse

VortexModel.Vortex.position